+++ packages/spor-react/src/theme/slot-recipes/alert-service.ts
@@ -35,7 +35,7 @@ export const alertServiceSlotRecipe = defineSlotRecipe({
       },
 
       _icon: {
-        color: "darkGrey",
+        color: "text",
       },
     },
     itemTriggerTitle: {
@@ -47,13 +47,13 @@ export const alertServiceSlotRecipe = defineSlotRecipe({
       textWrap: "nowrap",
     },
     itemBody: {
-      paddingInline: 2,
+      marginX: "auto",
+      padding: "0 !important",
       paddingBottom: ["0.5", null, null, "1"],
       color: "text.inverted",
       "& > p": {
         gap: 2,
         width: "full",
-        justifyContent: "center",
         borderBottom: "1px dashed",
         borderColor: "outline.inverted",
         paddingBottom: "3",