+++ packages/spor-react/src/theme/slot-recipes/radio-card.ts
@@ -45,7 +45,6 @@ export const radioCardSlotRecipe = defineSlotRecipe({
           outlineColor: "core.outline",
           outlineWidth: tokens.size.stroke.sm,
           outlineStyle: "solid",
-          backgroundColor: "core.surface",
 
           _hover: {
             outlineColor: "core.outline.hover",